requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.5.0
click==8.3.0
dotenv==0.9.9
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
python-dotenv==1.1.1
Werkzeug==3.1.3
WTForms==3.2.1
